consolide/reorganize/rename WT3 build & run files
