+++ watertap3/watertap3/utils/__init__.py
@@ -1,10 +1,5 @@
-
-from . import constituent_removal_water_recovery
-from .constituent_removal_water_recovery import *
 from . import cost_curves
 from .cost_curves import *
-from . import design
-from .design import *
 from . import financials
 from .financials import *
 from . import mixer_wt3
@@ -19,22 +14,18 @@ from . import splitter_wt3
 from .splitter_wt3 import *
 from . import splitter_binary
 from .splitter_binary import *
-from . import case_study_trains
-from .case_study_trains import *
-from . import generate_constituent_list
-from .generate_constituent_list import *
+from . import build_watertap3
+from .build_watertap3 import *
+from . import run_watertap3
+from .run_watertap3 import *
 from . import water_props
 from .water_props import *
-from . import watertap
-from .watertap import *
 from . import sensitivity_runs
 from .sensitivity_runs import *
 
 
 __all__ = [
-           *constituent_removal_water_recovery.__all__,
            *cost_curves.__all__,
-           *design.__all__,
            *financials.__all__,
            *mixer_wt3.__all__,
            *ml_regression.__all__,
@@ -43,8 +34,7 @@ __all__ = [
            *sensitivity_runs.__all__,
            *splitter_wt3.__all__,
            *splitter_binary.__all__,
-           *case_study_trains.__all__,
-           *generate_constituent_list.__all__,
            *water_props.__all__,
-           *watertap.__all__
+           *build_watertap3.__all__,
+           *run_watertap3.__all__
            ]
