+++ watertap3/watertap3/utils/water_props.py
@@ -3,7 +3,9 @@ from idaes.core.components import Component
 from idaes.core.phases import LiquidPhase
 from pyomo.environ import units as pyunits
 
-from . import generate_constituent_list
+# from . import generate_constituent_list
+import pandas as pd
+import numpy as np
 
 __all__ = ['WaterParameterBlock',
            'WaterStateBlock']
@@ -26,10 +28,25 @@ class PhysicalParameterData(PhysicalParameterBlock):
         self._state_block_class = WaterStateBlock
 
         self.Liq = LiquidPhase()
-        train_constituent_list = generate_constituent_list.run(self.parent_block())
+        # train_constituent_list = self.generate_constituent_list()
+        self.generate_constituent_list()
 
-        for constituent_name in train_constituent_list:
+        for constituent_name in self.train_constituent_list:
             setattr(self, constituent_name, Component())
+
+    def generate_constituent_list(self):
+        train = self.parent_block().train
+        # getting the list of consituents with removal factors that are bigger than 0
+        df = pd.read_csv('data/constituent_removal_factors.csv')
+        df.case_study = np.where(df.case_study == 'default', train['case_study'], df.case_study)
+        df = df[df.reference == train['reference']]
+        df = df[df.case_study == train['case_study']]
+        df = df[df.scenario == 'baseline']
+        list1 = df[df.value >= 0].constituent.unique()
+        list2 = self.parent_block().source_df.index.unique().to_list()
+
+        self.parent_block().source_constituents = source_constituents = [x for x in list1 if x in list2]
+        self.train_constituent_list = source_constituents
 
     @classmethod
     def define_metadata(cls, obj):
@@ -41,7 +58,6 @@ class PhysicalParameterData(PhysicalParameterBlock):
                 'temperature': pyunits.K,
                 'volume': pyunits.liter
                 })
-
 
 @declare_process_block_class("WaterStateBlock")
 class WaterStateBlockData(StateBlockData):
